Add files via upload
In this file, I demonstrate writing an incident report, while including DNS Network protocol is involved. I use the DNS and HTTP traffic logs to report on  port number, IP addresses, and time stamps. In this case the structure of the report, identifies the protocol involved in the incident, documents critical information involved in the incident, and concludes with at least one remediation for the incident, in this case a Brute Force Attack.