Add files via upload
The protocol impacted in the incident is Hypertext transfer protocol (HTTP). Running tcpdump and accessing the yummyrecipesforme.com website to detect the problem, capture protocol, and traffic activity in a DNS & HTTP traffic log file provided the evidence needed to come to this conclusion. The malicious file is observed being transported to the users’ computers using the HTTP protocol at the application layer. IT support recommended with immediate effect that we establish a policy to immediately change default.  Passwords are required to be 12 characters long, and must include lowercase, uppercase,  a minimum of two numbers and two special characters  i.e. *!@#$%&.